refactor(engine): restore cli and serve hygiene walls
Split the oversized test suites at cohesive behavior boundaries.

Move resident execution and lifecycle composition into nika-serve. This keeps nika-cli as a thin adapter without changing scheduling.

Tests: cargo test -p nika-cli -p nika-serve --lib

Tests: strict clippy, file and crate LOC gates, formatting, and diff checks
