+++ crates/nika-serve/Cargo.toml
@@ -20,6 +20,7 @@ miette     = { workspace = true, features = ["fancy-no-backtrace"] }
 nika-cadence = { path = "../nika-cadence", version = "0.115.0" }
 nika-error  = { path = "../nika-error", version = "0.115.0" }
 nika-execution = { path = "../nika-execution", version = "0.115.0" }
+nika-service-execution = { path = "../nika-service-execution", version = "0.115.0" }
 nika-fs    = { path = "../nika-fs", version = "0.115.0" }
 nika-runtime = { path = "../nika-runtime", version = "0.115.0" }
 nika-vocab = { path = "../nika-vocab", version = "0.115.0" }
