+++ crates/nika-serve/src/lib.rs
@@ -26,6 +26,8 @@ pub use schedule::{
 pub use server::{
     BoundServer, CredentialRefuse, ExecutionBackend, ExecutionDisposition, ExecutionOutcome,
     PreparedScheduledRun, ResidentAuthority, ResidentClock, ResidentConfig,
-    ResidentExecutionCoordinator, ServerConfig, ServerError, ServerLimits, SystemResidentClock,
-    serve_http,
+    ResidentExecutionBackend, ResidentExecutionCoordinator, ServerConfig, ServerError,
+    ServerLaunchRefuse, ServerLimits, SystemResidentClock, launch_operator_message,
+    optional_server_config, process_shutdown, serve_http, serve_resident, serve_resident_process,
+    server_operator_message,
 };
